This pull request addresses the following issue:
It is very hard for open source projects
to maintain current lists of the best  videos.
There are many Python conferences all over the world.  The best
videos keep changing.  One has to regularly call the YouTube
api's and sort the results to maintain the list.
One has to delete the old videos
when the products change sinificantly.

Fortunately PythonLinks.info/cython does all of thsi for you.
But that does not help the users browsing your site.

So I added an iFrame widget which displays the best Cython videos
organized into three different categories. As I index new conferences,
the changes will be displayed in this widget on your home page.

Thanks
Chris